requirements.txt
h5py
nltk
numpy
parse
python-gflags
sklearn
tqdm